requirements.txt
youtube_dl
PyNaCl
discord.py